adding orders

+++ Sixerr/Sixerr/Controllers/ProfilesController.cs
@@ -24,7 +24,7 @@ namespace Sixerr.Controllers
 
         public IActionResult MyProfile()
         {
-            uint id_ = _context.Profiles
+            long id_ = _context.Profiles
                 .First(p => p.User.Id == userManager.GetUserId(HttpContext.User)).Id;
             return RedirectToAction("Details", new { id = id_ });
         }
